app: rewrite smart-label to be calculated on widget's build-time
in the previous implementation, the ending of label was calculated
when proteinsPresentation() method was called.
It is invalid approach, since the widget built by this method has
was created with this ending already specified. But what, if value
of a.foundProteins has been changed before widget's build, but just
after its creation? Then, during this particular frame, the ending'd
be invalid!

Hehe - this is invisible for human viewer - I'm doing it just to be
honest :smile: